Add zero-hits exclusion rule for 'local x = nil'
'local x = nil' can run without triggering the debug hook in two cases:
the first, if under Lua 5.1 and it's the first statement in the function
(no corresponding bytecode instruction);
the second, under any Lua if it's not the first such statement in a row
(all assignments will be performed with one LOADNIL).

Fixes #41.