Convert statsfile and reportfile paths to absolute on config load
If a program using LuaCov changes its directory, particularly if it uses
`luacov.tick`, stats may be partially saved relatively to one of
intermediate current directories instead of the starting one.

The fix is to convert paths in config to absolute on load.
`pwd` (`cd` on Windows) is used to get current directory, so, no
dependency on lfs is introduced.

Older versions of LuaCov opened statsfile on config load, which avoids this
problem.