Remove './' prefix from the source filename.
The source filename will contain a './' prefix if they are run like
`lua -lluacov ./x.lua`.

As a result, if we run `lua -lluacov x.lua`, the `luacov.stats.out` is:
1:x.lua
1

However, if we run `lua -lluacov ./x.lua`, the `luacov.stats.out` is:
1:./x.lua
1

So the result could be inconsistent even we run in the same root directory of
the same project, which bring extra challenge to develop a custom reporter.